finished getEventiByOrganizzatore turistaPage

+++ BE/src/main/java/it/polimi/iswpf/dto/response/OrganizzatoreResponse.java
@@ -17,5 +17,5 @@ public class OrganizzatoreResponse {
 
     private LocalDateTime dataCreazioneAccount;
 
-    private Integer numeroEventiOrganizzati;
+    private Long numeroEventiOrganizzati;
 }
